+++ packages/rx_bloc_cli/bin/compile.sh
@@ -16,7 +16,7 @@ dart run rx_bloc_cli create \
   --enable-patrol \
   --realtime-communication=sse \
   --enable-otp \
-  --cicd=fastlane \
+  --cicd=github \
   --enable-pin-code \
   --no-interactive \
   example/test_app