+++ packages/rx_bloc_cli/lib/src/models/generator_arguments_provider.dart
@@ -133,6 +133,7 @@ class GeneratorArgumentsProvider {
     // CI/CD
     final cicdType = _reader.read<CICDType>(CommandArguments.cicd);
     final cicdEnabled = cicdType != CICDType.none;
+    final cicdGithubEnabled = cicdType == CICDType.github;
 
     return FeatureConfiguration(
       changeLanguageEnabled: changeLanguageEnabled,
@@ -145,6 +146,7 @@ class GeneratorArgumentsProvider {
       devMenuEnabled: devMenuEnabled,
       patrolTestsEnabled: patrolTestsEnabled,
       cicdEnabled: cicdEnabled,
+      cicdGithubEnabled: cicdGithubEnabled,
     );
   }