+++ packages/rx_bloc_cli/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [3.3.0]
+* Added Github as a CI/CD option under the flag `--cicd` (`fastlane` (default), `github`, none`)
+
 ## [3.1.2]
 * Update generated project dependencies
 * Fix OTP screen not showing on login
@@ -6,7 +9,7 @@
 * Fix generated project not properly building when `pin_code` is disabled
 
 ## [3.1.0]
-* Added Fastlane as a CI/CD option under the flag `--cicd` (fastlane` (default), `none`)
+* Added Fastlane as a CI/CD option under the flag `--cicd` (`fastlane` (default), `none`)
 * Added PIN code as an option under the flag `--enable-pin-code`
 * Introduce two new flavors (`sit` and `uat`) and remove `staging` flavor